requirements.txt
# 喵姆 AI 股市偵測站 - Python 套件需求
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
finmind==1.9.4
openpyxl>=3.1.0
pyngrok>=7.0.0

tqdm>=4.0.0
flask==3.0.0
yfinance>=0.2.0
nltk>=3.8.0
backtrader>=1.9.76.123
